Рефакторит функции изменения масштаба
