+++ js/main.js
@@ -1,3 +1,4 @@
 import './galery.js';
 import './img-upload-form.js';
-import './effects-slider';
+import './effects-slider.js';
+import './scale-controlls.js';
